+++ src/functions.php
@@ -244,6 +244,65 @@ function inOrderSequence($events)
 }
 
 /**
+ * Checks that at least one event is supplied.
+ *
+ * @param EventCollectionInterface $events,... The events.
+ *
+ * @return EventCollectionInterface|null The result.
+ * @throws InvalidArgumentException      If invalid input is supplied.
+ */
+function checkAnyOrder()
+{
+    return FacadeDriver::instance()->eventOrderVerifier()
+        ->checkAnyOrderSequence(func_get_args());
+}
+
+/**
+ * Throws an exception unless at least one event is supplied.
+ *
+ * @param EventCollectionInterface $events,... The events.
+ *
+ * @return EventCollectionInterface The result.
+ * @throws InvalidArgumentException If invalid input is supplied.
+ * @throws Exception                If the assertion fails, and the assertion recorder throws exceptions.
+ */
+function anyOrder()
+{
+    return FacadeDriver::instance()->eventOrderVerifier()
+        ->anyOrderSequence(func_get_args());
+}
+
+/**
+ * Checks if the supplied event sequence contains at least one event.
+ *
+ * @param mixed<EventCollectionInterface> $events The event sequence.
+ *
+ * @return EventCollectionInterface|null The result.
+ * @throws InvalidArgumentException      If invalid input is supplied.
+ */
+function checkAnyOrderSequence($events)
+{
+    return FacadeDriver::instance()->eventOrderVerifier()
+        ->checkAnyOrderSequence($events);
+}
+
+/**
+ * Throws an exception unless the supplied event sequence contains at least
+ * one event.
+ *
+ * @param mixed<EventCollectionInterface> $events The event sequence.
+ *
+ * @return EventCollectionInterface The result.
+ * @throws InvalidArgumentException If invalid input is supplied.
+ * @throws Exception                If the assertion fails, and the assertion recorder throws exceptions.
+ */
+function anyOrderSequence($events)
+{
+    return FacadeDriver::instance()->eventOrderVerifier()
+        ->anyOrderSequence($events);
+}
+
+/**
  * Create a new matcher that matches anything.
  *
  * @return MatcherInterface The newly created matcher.